% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-with_parameters.R
\name{with_parameters_test_that}
\alias{with_parameters_test_that}
\title{Legacy API for parameterized tests}
\usage{
with_parameters_test_that(
  desc_stub,
  code,
  ...,
  .cases = NULL,
  .test_name = NULL,
  .interpret_glue = TRUE
)
}
\arguments{
\item{desc_stub}{A string scalar. Used in creating the names of the
parameterized tests.}

\item{code}{Test code containing expectations.}

\item{...}{Named arguments of test parameters. All vectors should have the
same length.}

\item{.cases}{A data frame where each row contains test parameters.}

\item{.test_name}{An alternative way for providing test names. If provided,
the name will be appended to the stub description in `desc_stub`. If not
provided, test names will be automatically generated.}

\item{.interpret_glue}{Logical, default `TRUE`. If `FALSE`, and glue-like
markup in `desc_stub` is ignored, otherwise [glue::glue_data()] is
attempted to produce a more complete test description.}
}
\description{
We are moving away from this function name in favor of [test_these()].
}
